Refactored ignore attribute
So that it is contained within the dependency injection namespace instead of inside a generic attributes namespace
