+++ src/Hero6.Engine/Hero6.Engine.projitems
@@ -9,7 +9,7 @@
     <Import_RootNamespace>Hero6.Engine</Import_RootNamespace>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)Attributes\IgnoreAttribute.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)Services\DependencyInjection\IgnoreAttribute.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\CampaignExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\ColorExtensions.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Extensions\IEnumerableExtensions.cs" />
