+++ src/Hero6.Engine/ModuleController/Controller.cs
@@ -7,7 +7,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using LateStartStudio.Hero6.Attributes;
 using LateStartStudio.Hero6.MonoGame.GameLoop;
 using LateStartStudio.Hero6.Services.DependencyInjection;
 using LateStartStudio.Hero6.Services.UserInterfaces.Input.Mouse;
@@ -119,9 +118,8 @@ namespace LateStartStudio.Hero6.ModuleController
 
         public override string ToString() => $"Controller: {Module.Name}";
 
-        protected IEnumerable<Type> FindModules<T>() where T : IModule => Module.GetType().Assembly.GetTypes()
-            .Where(t => t.BaseType == typeof(T))
-            .Where(t => t.GetCustomAttributes(typeof(IgnoreAttribute), true).Length == 0);
+        protected IEnumerable<Type> FindModules<T>() where T : IModule =>
+            Module.GetType().Assembly.GetTypes().Where(t => t.BaseType == typeof(T));
 
         private void MouseMove(object sender, MouseMove e)
         {
